feat(admin v0)

+++ API/app/Http/Controllers/Admin/TypesController.php
@@ -7,9 +7,12 @@ use App\Http\Resources\TypesResource;
 use App\Models\Rtypes;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use App\Traits\HttpResponses;
 
 class TypesController extends Controller
 {
+    use HttpResponses;
     /**
      * Display a listing of the resource.
      */
@@ -27,10 +30,13 @@ class TypesController extends Controller
     {
         $validatedData = $request->validate([
             'name' => 'required|unique:rtypes|max:255',
+            'image' => ['file', 'mimes:jpeg,png,jpg,gif', 'max:4048'],
         ]);
 
+        $imagePath = $request->file('image')->store('public/images/types');
         $newType = Rtypes::create([
             'name' => $request->name,
+            'image' => str_replace('public/', 'storage/', $imagePath),
         ]);
 
         return new TypesResource($newType);
@@ -60,13 +66,54 @@ class TypesController extends Controller
 
         return new TypesResource($type);
     }
+    public function updateTypes(Request $request, $id)
+    {
+        $validatedData = $request->validate([
+            'name' => 'unique:rtypes|max:255',
+            'image' => ['file', 'mimes:jpeg,png,jpg,gif', 'max:4048'],
+        ]);
+        $Types = Rtypes::find($id);
+        if (!$Types) {
+            // If the real estate is not found, return a 404 response
+            return $this->error('', 'type not found', 404);
+        }
+        $updatedFields = [];
 
+        if ($request->file('image')) {
+            $image = $request->file('image');
+            $oldImagePath = str_replace('storage/', 'public/', $Types->image);
+
+            if ($oldImagePath && Storage::exists($oldImagePath)) {
+                // Delete the previous image if it exists
+                Storage::delete($oldImagePath);
+            }
+            // Store the new image and update the 'image' field in the updated fields array
+            $imagePath = $image->store('public/images/types');
+            $updatedFields['image'] = str_replace('public/', 'storage/', $imagePath);
+        }
+        if ($request->has('name')) {
+            $updatedFields['name'] = $request->name;
+        }
+        $Types->fill(array_filter($updatedFields));
+        $Types->save();
+        return new TypesResource($Types);
+    }
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Rtypes $type)
     {
-        $type->delete();
-        return response(null, 204);
+        if (!$type) {
+            return $this->error('', 'type not found', 404);
+        } else {
+            $imagePath = str_replace('storage/', 'public/', $type->image);
+            Storage::delete($imagePath);
+            $type->delete();
+
+            return $this->success([
+                'message' => 'The type has been deleted successfully',
+            ]);
+
+        }
     }
 }
